Add AUTO_PLAN to determine number of tests in file
A lot of my test files take the form of:

```c

void test_literals(void) {
    cmp_ok(3, "==", 3, "should find that 3 equals 3");
}

void test_arithmetic(void) {
    cmp_ok(1+1, "==", 2, "should find that 1+1 equals 2");
    cmp_ok(2*2, "==", 4, "should find that 2*2 equals 4");
}

int main(int argc, char **argv) {
    plan(NO_PLAN);
    test_literals();
    test_arithmetic();
    done_testing();
}
```

However, the problem with this is that it doesn't know the number of
tests in advance. The easy workaround is to run it once, determine the
number of tests, and then modify the `plan(NO_PLAN)` like to use that
number, but this is a little awkward, especially when the compiler could
do the counting for us.

----

`AUTO_PLAN` is an implementation of this idea of making the compiler
count for us. It uses the special preprocessor variable `__COUNTER__` to
count all of the lines that could produce an output line. The way to use
this feature is:

```c

void test_literals(void) {
    cmp_ok(3, "==", 3, "should find that 3 equals 3");
}

void test_arithmetic(void) {
    cmp_ok(1+1, "==", 2, "should find that 1+1 equals 2");
    cmp_ok(2*2, "==", 4, "should find that 2*2 equals 4");
}

int main(int argc, char **argv) {
    plan(AUTO_PLAN);
    test_literals();
    test_arithmetic();
    done_testing();
}
```

The compiler will automatically determine that it should plan for 3 test
cases (the 3 `cmp_ok` lines) and rewrite the `plan(AUTO_PLAN)` as
`plan(3)`. I decided to hide this feature behind an `#ifdef
USE_AUTO_PLAN` because there is a possibility that someone's code
already uses the `__COUNTER__` variable.

One of the major limitations of this feature is that you can't split
your tests up into multiple files and expect it to work, because the
`__COUNTER__` variable gets reset for each file. We could work around
this problem by using Boost, but that's too large of a dependency for a
simple library like this one.

Another limitation is that it will get the wrong count if you use
`pass()` and `fail()` separately. This is because they each use `ok()`
internally, but each `ok()` increases the number of tests, so if you had
a construct like:

```c
void test_pass_fail(void) {
    if (some_complicated_function()) {
        pass("Hooray");
    } else {
        fail("Boo");
    }
}
```

then this will be incorrectly counted as 2 tests instead of 1 (the
`pass()` gets transformed into `ok(1, "", "Hooray")` and similarly for
`fail()`). A workaround would be:

```c
void test_pass_fail(void) {
    int rv;

    rv = some_complicated_function();
    cmp_ok(rv, "==", 0, "should do something correctly");
}
```

which will be correctly counted as 1 test.

----

Let me know if you have any questions or concerns with this feature.